requirements.txt
# Core Dependencies
openai>=1.3.0
chromadb>=0.4.15
python-dotenv>=1.0.0

# Web Framework
flask>=3.0.0

# HuggingFace Text-to-Speech
transformers>=4.35.0
torch>=2.0.0
soundfile>=0.12.1
scipy>=1.11.0

# Additional utilities
numpy>=1.24.0
colorama>=0.4.6
